pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "google-adk>=1.20.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-text-splitters>=1.0.0",
    "ollama>=0.6.1",
    "pypdf>=6.4.0",
]
